Add new steps (setup pages) to the app setup flow
