+++ client/web/src/enterprise/app/settings/local-repositories/LocalRepositoriesTab.module.scss
@@ -11,7 +11,6 @@
 
 .container {
     padding: 1rem;
-    min-height: 16rem;
 }
 
 .list {
