--- client/web/src/enterprise/app/setup/AppSetupSteps.module.scss
@@ -1,136 +1,0 @@
-.root {
-    display: flex;
-    width: 100%;
-    height: 100%;
-    align-items: stretch;
-}
-
-.description {
-    display: flex;
-    flex-direction: column;
-    width: 50%;
-    height: 100%;
-    padding: 4rem 3.25rem 0;
-    color: var(--white);
-
-    &-heading {
-        font-size: 2rem;
-        font-weight: normal;
-        color: var(--white);
-    }
-
-    &-text {
-        color: #dedede;
-        font-size: 1.25rem;
-    }
-
-    &-image {
-        margin-top: auto;
-        max-width: calc(100% + 6.5rem);
-        margin-left: -1rem;
-        margin-right: -3.25rem;
-    }
-
-    &-footer {
-        color: #dedede;
-        font-size: 1rem;
-    }
-}
-
-.actions {
-    width: 50%;
-    height: 100%;
-    padding: 4rem 3.25rem 0;
-
-    &--with-bg {
-        background-color: var(--body-bg);
-    }
-
-    &-card {
-        display: flex;
-        flex-direction: column;
-        gap: 1.25rem;
-
-        padding: 2rem;
-        border: 1px solid #785198;
-        border-radius: 3px;
-        color: var(--white);
-        background: rgba(255, 255, 255, 0.1);
-    }
-
-    &-heading {
-        margin-bottom: 0;
-    }
-
-    &-buttons {
-        width: 100%;
-        display: flex;
-        flex-direction: column;
-        align-items: stretch;
-        gap: 0.75rem;
-    }
-
-    &-button {
-        display: flex;
-        align-items: center;
-        gap: 0.25rem;
-        justify-content: center;
-    }
-
-    &-footer {
-        margin-bottom: 0;
-    }
-
-    &-term-link {
-        color: var(--white);
-        text-decoration: underline;
-    }
-}
-
-.local-repositories {
-    &-navigation {
-        display: flex;
-        gap: 0.5rem;
-        margin-top: 1.5rem;
-    }
-
-    &-buttons-group {
-        justify-content: stretch;
-        align-items: center;
-        min-width: 0;
-
-        > * {
-            flex-grow: 1;
-            flex-shrink: 0;
-        }
-    }
-
-    &-widget {
-        margin: -2rem -1.5rem;
-        display: flex;
-        flex-direction: column;
-        gap: 0.75rem;
-        width: auto;
-    }
-}
-
-.extension {
-    display: flex;
-    flex-direction: column;
-    margin: auto;
-    padding: 0 3rem;
-    width: 100%;
-
-    &-heading {
-        margin-bottom: 0.5rem;
-        font-size: 2rem;
-        font-weight: normal;
-        color: var(--white);
-    }
-
-    &--footer {
-        display: flex;
-        margin-top: 1.5rem;
-        gap: 0.5rem;
-    }
-}
