+++ client/web/src/enterprise/app/setup/AppSetupWizard.module.scss
@@ -1,9 +1,6 @@
-.root {
-    width: 100%;
-    height: 100%;
-}
-
 .content {
-    height: 100%;
     background: radial-gradient(116.94% 133.26% at 52.08% 100%, #6112a3 0.41%, #270741 54.04%, #1e0632 94.78%);
+
+    min-height: 100vh;
+    align-self: baseline;
 }
