+++ client/web/src/enterprise/app/setup/AppSetupWizard.tsx
@@ -6,7 +6,10 @@ import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryServi
 import { PageTitle } from '../../../components/PageTitle'
 import { StepConfiguration, SetupStepsContent, SetupStepsRoot } from '../../../setup-wizard'
 
-import { AppWelcomeSetupPage, AddLocalRepositoriesSetupPage, InstallExtensionsSetupPage } from './AppSetupSteps'
+import { AppAllSetSetupStep } from './steps/AppAllSetSetupStep'
+import { AppInstallExtensionsSetupStep } from './steps/AppInstallExtensionsSetupStep'
+import { AddLocalRepositoriesSetupPage } from './steps/AppLocalRepositoriesSetupStep'
+import { AppWelcomeSetupStep } from './steps/AppWelcomeSetupStep'
 
 import styles from './AppSetupWizard.module.scss'
 
@@ -15,7 +18,7 @@ const APP_SETUP_STEPS: StepConfiguration[] = [
         id: 'welcome',
         name: 'Welcome page',
         path: 'welcome',
-        component: AppWelcomeSetupPage,
+        component: AppWelcomeSetupStep,
     },
     {
         id: 'local-repositories',
@@ -27,7 +30,13 @@ const APP_SETUP_STEPS: StepConfiguration[] = [
         id: 'install-extensions',
         name: 'Install Sourcegraph extensions',
         path: 'install-extensions',
-        component: InstallExtensionsSetupPage,
+        component: AppInstallExtensionsSetupStep,
+    },
+    {
+        id: 'all-set',
+        name: 'All set',
+        path: 'all-set',
+        component: AppAllSetSetupStep,
         nextURL: '/',
         onView: () => {
             localStorage.setItem('app.setup.finished', 'true')
@@ -54,7 +63,7 @@ export const AppSetupWizard: FC<TelemetryProps> = ({ telemetryService }) => {
     }
 
     return (
-        <div className={styles.root}>
+        <>
             <PageTitle title="Sourcegraph App setup" />
 
             <SetupStepsRoot
@@ -65,6 +74,6 @@ export const AppSetupWizard: FC<TelemetryProps> = ({ telemetryService }) => {
             >
                 <SetupStepsContent telemetryService={telemetryService} className={styles.content} />
             </SetupStepsRoot>
-        </div>
+        </>
     )
 }
