+++ client/web/src/enterprise/app/setup/steps/AppWelcomeSetupStep.module.scss
@@ -1,0 +1,97 @@
+@import 'wildcard/src/global-styles/breakpoints';
+
+.root {
+    display: flex;
+    flex-direction: column;
+    padding: 4.5rem 3.5rem 0;
+
+    // Tablet and "mobile" (small) layout
+    @media (--md-breakpoint-down) {
+        .content {
+            grid-template-areas:
+                'description description'
+                'actions actions';
+            margin-bottom: 2.5rem;
+        }
+
+        .image {
+            max-width: 100%;
+            margin: auto 0 0 0;
+        }
+    }
+}
+
+.content {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    row-gap: 1.5rem;
+    column-gap: 3rem;
+    grid-template-areas: 'description actions';
+}
+
+.description {
+    grid-area: description;
+    display: flex;
+    flex-direction: column;
+
+    &-heading {
+        font-size: 2rem;
+        font-weight: normal;
+        color: var(--white);
+    }
+
+    &-text {
+        color: #dedede;
+        font-size: 1.25rem;
+    }
+
+    &-footer {
+        color: #dedede;
+        font-size: 1rem;
+    }
+}
+
+.actions {
+    grid-area: actions;
+
+    &-card {
+        display: flex;
+        flex-direction: column;
+        gap: 1.25rem;
+        padding: 2rem;
+        border-radius: 3px;
+        border: 1px solid #785198;
+        color: var(--white);
+        background: rgba(255, 255, 255, 0.1);
+    }
+
+    &-buttons-group {
+        width: 100%;
+        display: flex;
+        flex-direction: column;
+        align-items: stretch;
+        gap: 0.75rem;
+    }
+
+    &-button {
+        display: flex;
+        gap: 0.25rem;
+        align-items: center;
+        justify-content: center;
+    }
+
+    &-footer {
+        margin-bottom: 0;
+    }
+
+    &-term-link {
+        color: var(--white);
+        text-decoration: underline;
+    }
+}
+
+.image {
+    max-width: 65%;
+    margin-top: auto;
+    margin-left: -3.5rem;
+}
